# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/person/": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Retrieve all people from the database",
        "description": "This endpoint fetches a list of all people in the database.",
        "responses": {
          "200": {
            "description": "List of all people.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                  },
                  "username": {
                    "type": "string",
                    "example": "johndoe123"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "locationId": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "Software Engineer"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "person"
        ],
        "summary": "Add a person to the database",
        "description": "This endpoint adds a new person to the database. The request body must comply with the validation schema.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "The details of the person to add.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe123"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+1234567890"
                },
                "locationId": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Software Engineer"
                }
              },
              "required": [
                "firstName",
                "lastName",
                "email",
                "username",
                "phoneNumber",
                "locationId",
                "title"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person successfully added and token generated.",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "token": {
                  "type": "string",
                  "example": "<jwt-token>"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Validation or database error occurred.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized. Validation failed or a required field is missing.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/first-name/{firstName}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get people by first name",
        "description": "This endpoint retrieves all people with the specified first name.",
        "parameters": [
          {
            "name": "firstName",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The first name to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "List of people with the specified first name.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                  },
                  "username": {
                    "type": "string",
                    "example": "johndoe123"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "locationId": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "Software Engineer"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No people found with the specified first name.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such person with first name: John"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/last-name/{lastName}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get people by last name",
        "description": "This endpoint retrieves all people with the specified last name.",
        "parameters": [
          {
            "name": "lastName",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The last name to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "List of people with the specified last name.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                  },
                  "username": {
                    "type": "string",
                    "example": "johndoe123"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "locationId": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "Software Engineer"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No people found with the specified last name.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such person with last name: Doe"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/email/{email}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get a person by email",
        "description": "This endpoint retrieves a person with the specified email address.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The email address to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "Person with the specified email address.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe123"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+1234567890"
                },
                "locationId": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Software Engineer"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No person found with the specified email address.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such person with email: johndoe@example.com"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/username/{username}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get a person by username",
        "description": "This endpoint retrieves a person with the specified username.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The username to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "Person with the specified username.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe123"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+1234567890"
                },
                "locationId": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Software Engineer"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No person found with the specified username.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such person with username: johndoe123"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/phone-number/{phoneNumber}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get a person by phone number",
        "description": "This endpoint retrieves a person with the specified phone number.",
        "parameters": [
          {
            "name": "phoneNumber",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The phone number to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "Person with the specified phone number.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe123"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+1234567890"
                },
                "locationId": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Software Engineer"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No person found with the specified phone number.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such person with phone number: +1234567890"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/me": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get details of the authenticated user",
        "description": "This endpoint retrieves the details of the user associated with the provided token.",
        "responses": {
          "200": {
            "description": "Authenticated user details.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe123"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+1234567890"
                },
                "locationId": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Software Engineer"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Error in token decoding or database lookup.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/person/{id}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Get a person by their unique ID",
        "description": "This endpoint retrieves a person with the specified unique ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "The unique ID of the person to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Person with the specified unique ID.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe123"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+1234567890"
                },
                "locationId": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Software Engineer"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No person found with the specified unique ID.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such person with id: 1"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}